+++ primality_deter/euclide.hpp
@@ -2,8 +2,8 @@
 #define euclide_hpp
 
 
-unsigned int pgcd(unsigned int a, unsigned int b);
-bool computation_bound_euclide(unsigned int number);
-bool computation_bound_modulo(unsigned int number);
+unsigned long int pgcd(unsigned long int a, unsigned long int b);
+bool computation_bound_euclide(unsigned long int number);
+bool computation_bound_modulo(unsigned long int number);
 
 #endif
