+++ primality_proba/millerRabin.hpp
@@ -5,8 +5,8 @@
 #include <cstdlib>
 #include <ctime>
 
-unsigned int expPower(unsigned int x, unsigned int y, unsigned int p);
-bool millerTest(unsigned int d, unsigned int n);
-bool millerRabin(unsigned int n, unsigned int k);
+unsigned long int expPower(unsigned long int x, unsigned long int y, unsigned long int p);
+bool millerTest(unsigned long int d, unsigned long int n);
+bool millerRabin(unsigned long int n, unsigned long int k);
 
 #endif
